refactor(core): Make Peeker not a Printer
Description
---
Moves Peeker into its own file and makes it not a Printer. This allows me to make Printer non-generic and add message deduplication functionality which is necessary for filestream error message reporting.

Reviewers: kptkin

Reviewed By: kptkin

Pull Request: https://github.com/wandb/wandb/pull/7558
