+++ core/pkg/server/stream.go
@@ -151,11 +151,11 @@ func NewStream(settings *settings.Settings, _ string) *Stream {
 	})
 
 	// TODO: replace this with a logger that can be read by the user
-	peeker := observability.NewPeeker()
+	peeker := &observability.Peeker{}
+	terminalPrinter := observability.NewPrinter()
 
 	backendOrNil := NewBackend(s.logger, settings)
 	fileTransferStats := filetransfer.NewFileTransferStats()
-	terminalPrinter := observability.NewPrinter[string]()
 	var graphqlClientOrNil graphql.Client
 	var fileStreamOrNil filestream.FileStream
 	var fileTransferManagerOrNil filetransfer.FileTransferManager
